reduce width of onsen_id.

+++ backend/backend.sql
@@ -42,7 +42,7 @@ INSERT INTO DOUBUTSU.game_seq VALUES (0);
 
 -- * 温泉 (ステージ?) - 属性
 CREATE TABLE DOUBUTSU.onsen
-  ( id INTEGER NOT NULL
+  ( id SMALLINT NOT NULL
   , onsen_name TEXT NOT NULL
   , released_at TIMESTAMP NOT NULL
 
@@ -50,14 +50,14 @@ CREATE TABLE DOUBUTSU.onsen
   );
 
 CREATE TABLE DOUBUTSU.onsen_seq
-  ( seq INTEGER NOT NULL
+  ( seq SMALLINT NOT NULL
   );
 
 INSERT INTO DOUBUTSU.onsen_seq VALUES (0);
 
 -- * スロット (動物が温泉に入ったり、カスタマイズ用の木の種や実を植えたりする) - 属性
 CREATE TABLE DOUBUTSU.slot
-  ( onsen_id INTEGER NOT NULL
+  ( onsen_id SMALLINT NOT NULL
   , local_slot_number INTEGER NOT NULL
   , locate_x INTEGER NOT NULL
   , locate_y INTEGER NOT NULL
@@ -126,7 +126,7 @@ CREATE TABLE DOUBUTSU.item_stock
 
 -- * ミッション (温泉レベルアップ条件) - 属性
 CREATE TABLE DOUBUTSU.mission
-  ( onsen_id INTEGER NOT NULL
+  ( onsen_id SMALLINT NOT NULL
   , onsen_level INTEGER NOT NULL
   , mission_type INTEGER NOT NULL
 
@@ -137,7 +137,7 @@ CREATE TABLE DOUBUTSU.mission
 -- ** はじめてのお客さん - 属性
 -- mission_type == 0
 CREATE TABLE DOUBUTSU.mission_first_visited
-  ( onsen_id INTEGER NOT NULL
+  ( onsen_id SMALLINT NOT NULL
   , onsen_level INTEGER NOT NULL
   , doubutsu_id SMALLINT NOT NULL
 
@@ -147,7 +147,7 @@ CREATE TABLE DOUBUTSU.mission_first_visited
 -- ** 特定の動物がn回、来た - 属性
 -- mission_type == 1
 CREATE TABLE DOUBUTSU.mission_visited
-  ( onsen_id INTEGER NOT NULL
+  ( onsen_id SMALLINT NOT NULL
   , onsen_level INTEGER NOT NULL
   , doubutsu_id SMALLINT NOT NULL
   , times INTEGER NOT NULL
@@ -158,7 +158,7 @@ CREATE TABLE DOUBUTSU.mission_visited
 -- ** 特定の種あるいは木の実をn回、植えた - 属性
 -- mission_type == 2
 CREATE TABLE DOUBUTSU.mission_used_item
-  ( onsen_id INTEGER NOT NULL
+  ( onsen_id SMALLINT NOT NULL
   , onsen_level INTEGER NOT NULL
   , item_id INTEGER NOT NULL
   , times INTEGER NOT NULL
@@ -174,7 +174,7 @@ CREATE TABLE DOUBUTSU.mission_used_item
 CREATE TABLE DOUBUTSU.event_log
   ( id BIGINT NOT NULL
   , game_id INTEGER NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , event_type INTEGER NOT NULL
   , created_at TIMESTAMP NOT NULL
 
@@ -196,7 +196,7 @@ INSERT INTO DOUBUTSU.event_log_seq VALUES (0);
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_action
   ( event_log_id BIGINT NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , local_slot_number INTEGER NOT NULL
   , doubutsu_id SMALLINT NOT NULL
   , action_type INTEGER NOT NULL -- 来る: 0 , 帰る: 1 , or 癒しポーズID
@@ -221,7 +221,7 @@ CREATE TABLE DOUBUTSU.event_item
 --  event_data == id
 CREATE TABLE DOUBUTSU.event_growth
   ( event_log_id BIGINT NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , local_slot_number INTEGER NOT NULL
   , item_id INTEGER NOT NULL
   , growth_level INTEGER NOT NULL -- 植えたときはレベル1
@@ -258,7 +258,7 @@ CREATE TABLE DOUBUTSU.event_whether
 -- * 温泉の状態 - 属性
 CREATE TABLE DOUBUTSU.onsen_status
   ( game_id  INTEGER NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , updated_at TIMESTAMP NOT NULL
   , onsen_level INTEGER NOT NULL
   , mission_status INTEGER NOT NULL  -- 未開始: 0 , 実行中: 1 , 達成: 2
@@ -273,7 +273,7 @@ CREATE TABLE DOUBUTSU.onsen_status
 -- slot_type == 0
 CREATE TABLE DOUBUTSU.slot_status_doubutsu
   ( game_id  INTEGER NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , local_slot_number INTEGER NOT NULL
   , doubutsu_id SMALLINT NOT NULL -- 入っていない: -1
   -- , doubutsu_pause INTEGER NOT NULL
@@ -285,7 +285,7 @@ CREATE TABLE DOUBUTSU.slot_status_doubutsu
 -- slot_type == 1
 CREATE TABLE DOUBUTSU.slot_status_custom
   ( game_id  INTEGER NOT NULL
-  , onsen_id INTEGER NOT NULL
+  , onsen_id SMALLINT NOT NULL
   , local_slot_number INTEGER NOT NULL
   , item_id INTEGER NOT NULL -- 何も植えられていない: -1
   , growth_level INTEGER NOT NULL
